[RFC] Issue tracking hooks using Gerrit 2.6 commit validation
Has a commonreusable infrastructure to build the issue-tracking
association logic to other Issue-Trackers (i.e. BugZilla, JIRA).
The generic issue-tracker association plugin is hooks-its and
provides the logic for:
> Insertion of comment links using regex patterns
> Enforcement of issue-ids in git commits using
  Gerrit 2.6 commit validation listeners
> Automation of issue-tracker status transition
  based on Gerrit code-review status
hooks-its does not make assumption on the underlying issue-tracker
system but provides only an abstract facade interface ItsFacade to
access and perform actions on the issue-tracker.

Change-Id: I82ceb9a5eae825e6216121365f15ed4d4d59acc4