Update patch set 1
Patch Set 1:

The simpler the code hierarchy, the easier the code is to understand.


> a) hooks-its-parent can be used to share the common dependencies
> needed for all the hooks-its-X plugins (Jira, Bugzilla or others)

As this sounds over-engineered to me, I assume I do not understand
what you mean. So let me flesh out an example:

Lets consider we have the hooks-its project split into hooks-its,
hooks-its-parent artifacts, as it is now. And consider a dependency on
libX, which is /not/ needed by hooks-its, but is a common dependency of
hooks-bugzilla, hooks-jira, ....

If I understood you correctly, this would be the show case for
hooks-its-parent, as we can put the dependency on libX into
hooks-its-parent?

But that would mean that although hooks-bugzilla directly depends on
libX, this dependency is not visible directly from hooks-bugzilla. We'd
obfuscate this dependency and hide it away.

Consider I implement a further hooks package for a further ITS, e.g.:
Mantis. hooks-mantis does not depend on libX. So libX is no longer a
„common dependency” so we have to move the libX out of hooks-its and
finally into each of hooks-bugzilla, hooks-jira,...  anyways.

Sounds like unwarranted extra effort to me.


Ignoring those intangible „let's have simple code” issues, what would
be candidates for dependencies within hooks-its-parent?

My local version hooks-bugzilla with cleaned up dependencies depends
solely on hooks-its (without -parent) and j2bugzilla.
* hooks-its is probably a common dependency, but trading hooks-its by
  hooks-its-parent would not buy us anything -- besides obfuscation.
* hooks-its is probably a common dependency, but trading hooks-its by
  hooks-its-parent would not buy us anything -- besides obfuscation.
* j2bugzilla is specific to hooks-bugzilla, so it does not qualify as
  common dependency either.

So are we sacrificing simpleness of our code hierarchy for a non
existing use case?

> b) hooks-its now supports commit validation and workflow, but I was
> thinking about introducing backend groups for Issue Trackers Systems
> and other extensions: those would fit nicely in another Maven module
> within the hooks-its project.

What specifially do you mean by „backend groups”?
Could you provide an example?

After all we split off hooks-its, hooks-bugzilla from hooks-jira to
decouple things and allow for better, simpler reuse, didn't we?

Patch-set: 1