Update patch set 1
Patch Set 1:

Putting suggested common version of common dependencies on the parent project would enforce hooks-its related extensions to use them. 

Additionally common variables (i.e. pointing to a Gerrit API version spec) can be put into the parent module and do not need to be duplicated across different hooks-X projects.

NOTE: this does not mean that hooks-X and hooks-Y have to share the same dependencies but rather *IF* they happen to need the same library, the version of it can be defined in the parent project facilitating consistency.

With regards to the "Group backend" support, it is something we have in the pipeline to be contributed to the hooks-its BUT we did not want every hooks-X or hooks-Y to be forced to implement !
Gerrit "Group backend" is a pluggable external repository to retrieve groups, when dealing with external Issue Trackers (i.e. Jira) they could be a source of "existing roles" that could be reused in Gerrit.

This concept then is common to potentially every issue tracker (i.e. CollabNet TeamForge has the same logic) and thus could be located into hooks-its project but not necessarily in the current "commit validation and workflow" module. Even the current "commit validation and workflow" should be split into two maven modules IMHO, as the description implies two different purposes ;-)

I am opened to change my mind but it seems that splitting into separate Maven Modules would allow to limit the cross-dependencies when the implemented issue-tracker (i.e. Bugzilla) would not require the complete support for all the capabilities.

Luca.

Patch-set: 1