Update patch set 1
Patch Set 1: Code-Review-1

This will work if you've got one option set to true and the other to false, or both to false, but what if both are set to true? Wouldn't you get a redundant notification?

IMHO, we should probably just skip commentOnChangeCreated if commentOnPatchSetCreated is also true.

Patch-set: 1