Update patch set 1
Patch Set 1:

I thought I would submit this one for some comments. It helped me to build the hook-its plugin which I needed to build the hooks-jira plugin. A step further would be to deploy the artifact into a public repo so the other hooks-* plugins could be built without first requiring gerrit+hooks-its to be built and deployed locally.

Patch-set: 1
Label: Verified=0