Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Currently, for example the "hooks-bugzilla" plugin shows as "hooks-bugzilla" in gerrit, and is configured via a "bugzilla" commentLink.
That makes sense to me.

The plugin comes with a distinctive name, but for configuration, we use a name that's closer to the ITS.

After the change, to be configured via "bugzilla", the "hooks-bugzilla" has to come with the plugin-name "bugzilla", and hence would also show as "bugzilla" within gerrit. It's confusing to me that the "hooks-bugzilla" plugin will show up as "bugzilla".

While I see the elegance of reusing the plugin name, from my point of view showing "bugzilla" everywhere is confusing. Users already have a hard time to distinguish between plugins and products, and get confused. This change makes the distinction between the plugin and ITS even less visible.

Patch-set: 1
Label: Code-Review=-1