Update patch set 1
Patch Set 1:

First of all thanks for the comments, they are much appreciated!

 > Currently, for example the "hooks-bugzilla" plugin shows as
 > "hooks-bugzilla" in gerrit, and is configured via a "bugzilla"
 > commentLink. That makes sense to me.

I see your point and I actually agree to it. It's nice to have the
plugin named as "hooks-bugzilla" and have the configuration in a
"bugzilla" section. Maybe I should shortly explain the big picture of
what I'm trying to achieve, then it's easier to say why I proposed
"bugzilla" as the plugin name here.

Basically we have a Gerrit server with several thousands projects, but
only some teams want to use the integration with the issue tracker.
This is why we have the requirement to be able to enable/disable the
issue tracker integration per project. For us the project owners
should be able to make this decision and it should be as easy as
possible to switch the integration on/off. At best we would just have
a checkbox in the UI with which the issue tracker system can be
enabled/disabled for a project. Also we may have the need to support
integration with different issue tracker instances of the same type
(e.g. two installations of Jira).

On the other side I don't want to change any existing behaviour for
those that already use the plugin. This means:
1. it must be easy to enable the integration for all projects
2. it must be possible to enforce the integration for all projects
3. there shouldn't be any changes to the way how the plugin is
configured now

To enable/disable the issue tracker integration per project I want to
make use of the infrastructure from Gerrit core that allows to store
plugin configuration parameters in the projects project.config file.
This information is stored in a subsection which is named after the
plugin.

  [plugin "<plugin-name>"]
    myParam = myValue

I think it would be nice if the subsection here would have the same
name as the section in which the global plugin configuration is
stored:

  [plugin "bugzilla"]
    myParam = myValue

I would be more happy with having it as

  [plugin "hooks-bugzilla"]
    myParam = myValue

and then renaming the section in the global configuration to
"hooks-bugzilla" too, but this would mean to change the way of how the
plugin is configured now. This means admins on upgrade of the plugin
would need to adapt their configuration or an init step in the plugin
would need to do this reconfiguration automatically. This is why I
decided to rather use "bugzilla" as plugin name.

In the future we may even want to have different plugin configurations per
project, e.g. different teams may have different opinions about what
action should be done in the issue tracker on uploading a new patch
set, etc. The idea here is to allow in the project.config (or in an
own config file) to overwrite the global configuration, again it would
be nice to have the names aligned.

Especially when it comes to exposing the plugin configuration in the
UI, the project specific plugin configuration depends on the plugin
name (because this code is quite generic). 

As I said it's just nice to have this subsection having the same name
as the section in the global config, this alone is not enough
motivation for this change.

If we want to support different instances of the same issue tracker
type we need to have different global configuration for each instance.
If the section for the global configuration is named after the plugin
this gets quite easy, as we can just change the plugin name in the
manifest and deploy the same plugin multiple times with different
names. This is why the plugin name should be used for the global
configuration section and I've chosen "bugzilla" as plugin name to
stay compatible with existing configurations.

Of course there are some other possibilities to achieve the same, but
I found that this approach is the easiest.

Patch-set: 1
Reviewer: ekempin <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0