Update patch set 5
Patch Set 5: Code-Review-1

As we have the mechanism to enable/ disable on a per-branch basis, why do we want to make a special case for refs/meta/config ?

When refs/meta/config is updated from external git push, we may still want to ability to require a ticket for it (i.e. request to the SCM manager for performing that change).

SOX-compliant clients that actually requires that everything could be potentially forced to be associated to a ticket request, and this may potentially create a dangerous "backdoor".

If you want to avoid validating on refs/meta/config, why don't you just define a ref-ex for it ?

Patch-set: 5
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1