Update patch set 5
Patch Set 5:

> As we have the mechanism to enable/ disable on a per-branch basis,
 > why do we want to make a special case for refs/meta/config ?
 > 
 > When refs/meta/config is updated from external git push, we may
 > still want to ability to require a ticket for it (i.e. request to
 > the SCM manager for performing that change).
 > 
 > SOX-compliant clients that actually requires that everything could
 > be potentially forced to be associated to a ticket request, and
 > this may potentially create a dangerous "backdoor".
 > 
 > If you want to avoid validating on refs/meta/config, why don't you
 > just define a ref-ex for it ?

Yes, you are right since we are now having the possibility to configure the branches for the ITS integration, we can use this feature to exclude the refs/meta/config branch and we don't need any hard-coded logic for this.

I've added a new InitStep which allows to configure the branches for the ITS integration and by default it proposes refs/heads/* [1], but this can easily be changed to refs/* if the refs/meta/config branch should be included.

[1] https://gerrit-review.googlesource.com/53690

Patch-set: 5
Reviewer: ekempin <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0